Move section to folder, change name to styled.js
